Minor issues
Issue#1: Notification Bar » Elementor: bar not getting any style. Fully broken.

Issue#2: As we added the Link type to the "Contact Form" and "Email Subscription", we have to add the "Open Link In New Tab" option in the Customize » Behaviour tab. Currently, it is hidden.

Issue#3: Throwing PHP errors in "Inline Blocks" and "Shortcode".